+++ src/Blog.Services/Implementations/CommentService.cs
@@ -1,26 +1,31 @@
 ﻿using AutoMapper;
 using Blog.Bases.Services;
 using Blog.Entities.Comments;
+using Blog.Entities.Posts;
 using Blog.Repositories.Abstractions;
 using Blog.Repositories.Entities;
 using Blog.Services.Abstractions;
 using Blog.Services.Entities;
+using Blog.Translations.Abstractions;
 using Blog.Validations;
 using Blog.Validations.Abstractions;
+using Microsoft.AspNetCore.Http;
 
 namespace Blog.Services.Implementations
 {
-    public class CommentService : ICommentService
+    public class CommentService : ServiceBase, ICommentService
     {
         private readonly ICommentRepository _repository;
         private readonly IValidationFactory<CommentInput> _commentValidation;
         private readonly IValidationFactory<Guid> _idValidation;
         private readonly IMapper _mapper;
 
-        public CommentService(ICommentRepository repository, 
+        public CommentService(IHttpContextAccessor httpContextAccessor,
+            ITranslationResource translationResource,
+            ICommentRepository repository,
             IValidationFactory<CommentInput> postValidation,
             IValidationFactory<Guid> idValidation,
-            IMapper mapper)
+            IMapper mapper) : base(httpContextAccessor, translationResource)
         {
             _repository = repository;
             _commentValidation = postValidation;
@@ -35,10 +40,14 @@ namespace Blog.Services.Implementations
 
             if (validation.Success)
             {
-                RepositoryOutput<Comment> repositoryResult = await _repository.GetCommentAsync(id);
-                result.Message = repositoryResult.Message;
-                result.Output = _mapper.Map<CommentOutput>(repositoryResult.Output);
-                result.Errors = repositoryResult.Errors;
+                RepositoryOutput<Comment> internalComment = await _repository.GetInternalCommentAsync(id);
+                if (ValidateOwnerOrAdmin(internalComment.Output.CommentAuthorId, result))
+                {
+                    RepositoryOutput<Comment> repositoryResult = await _repository.GetCommentAsync(id);
+                    result.Message = repositoryResult.Message;
+                    result.Output = _mapper.Map<CommentOutput>(repositoryResult.Output);
+                    result.Errors = repositoryResult.Errors;
+                }
             }
             else
             {
@@ -53,7 +62,7 @@ namespace Blog.Services.Implementations
             ServiceOutput<CommentOutput> result = new();
             ValidationOutput validation = await _commentValidation.ValidateAsync(input.Input);
 
-            if (validation.Success)
+            if (validation.Success && ValidateOwnerOrAdmin(input.Input.CommentAuthorId, result))
             {
                 RepositoryOutput<Comment> repositoryResult = await _repository.CreateCommentAsync(new RepositoryInput<Comment>()
                 {
@@ -65,7 +74,7 @@ namespace Blog.Services.Implementations
             }
             else
             {
-                result.Errors = validation.Errors;
+                if (validation.Errors != null) { result.Errors = validation.Errors; }
             }
 
             return result;
@@ -75,7 +84,7 @@ namespace Blog.Services.Implementations
         {
             ServiceOutput<CommentOutput> result = new();
             ValidationOutput validation = await _commentValidation.ValidateAsync(input.Input);
-            if (validation.Success)
+            if (validation.Success && ValidateOwnerOrAdmin(input.Input.CommentAuthorId, result))
             {
                 RepositoryOutput<Comment> repositoryResult = await _repository.UpdateCommentAsync(new RepositoryInput<Comment>()
                 {
@@ -87,7 +96,7 @@ namespace Blog.Services.Implementations
             }
             else
             {
-                result.Errors = validation.Errors;
+                if (validation.Errors != null) { result.Errors = validation.Errors; }
             }
 
             return result;
@@ -100,10 +109,14 @@ namespace Blog.Services.Implementations
 
             if (validation.Success)
             {
-                RepositoryOutput<bool> repositoryResult = await _repository.RemoveCommentAsync(id);
-                result.Message = repositoryResult.Message;
-                result.Output = repositoryResult.Output;
-                result.Errors = repositoryResult.Errors;
+                RepositoryOutput<Comment> internalComment = await _repository.GetInternalCommentAsync(id);
+                if (ValidateOwnerOrAdmin(internalComment.Output.CommentAuthorId, result))
+                {
+                    RepositoryOutput<bool> repositoryResult = await _repository.RemoveCommentAsync(id);
+                    result.Message = repositoryResult.Message;
+                    result.Output = repositoryResult.Output;
+                    result.Errors = repositoryResult.Errors;
+                }
             }
             else
             {
